fix: restore parent collection when child page is restored

+++ pages/api/pages/[pageId].ts
@@ -70,28 +70,45 @@ export default async function pageHandler(
         select: pageSelectWithTextContent,
       });
 
-      // delete the page from typesense db on soft delete
-      // and add it back on restore
-      if (!isDeleted) {
-        const typesensePage: typesensePageDocument = {
-          id: updatedPage.id,
-          userId: updatedPage.userId,
-          pageName: updatedPage.pageName,
-          pageTextContent: updatedPage.textContent,
-          pageCreatedAt: updatedPage.createdAt.getTime(),
-          pageModifiedAt: updatedPage.modifiedAt.getTime(),
-          isFavourite: updatedPage.isFavourite,
-        };
+      // On restoring a page, restore its collection as well
+      if (typeof isDeleted !== "undefined" && !isDeleted) {
+        await prisma.collection.update({
+          where: {
+            id_userId: {
+              id: updatedPage.collectionId,
+              userId: session.accountId,
+            },
+          },
+          data: {
+            isDeleted: false,
+          },
+        });
+      }
 
-        await serverTypesenseClient
-          .collections("pages")
-          .documents()
-          .upsert(typesensePage);
-      } else {
-        await serverTypesenseClient
-          .collections("pages")
-          .documents(updatedPage.id)
-          .delete();
+      // Delete the page from typesense db on soft delete
+      // and add it back on restore
+      if (typeof isDeleted !== "undefined") {
+        if (!isDeleted) {
+          const typesensePage: typesensePageDocument = {
+            id: updatedPage.id,
+            userId: updatedPage.userId,
+            pageName: updatedPage.pageName,
+            pageTextContent: updatedPage.textContent,
+            pageCreatedAt: updatedPage.createdAt.getTime(),
+            pageModifiedAt: updatedPage.modifiedAt.getTime(),
+            isFavourite: updatedPage.isFavourite,
+          };
+
+          await serverTypesenseClient
+            .collections("pages")
+            .documents()
+            .upsert(typesensePage);
+        } else {
+          await serverTypesenseClient
+            .collections("pages")
+            .documents(updatedPage.id)
+            .delete();
+        }
       }
 
       // transform data for client
